+++ lib/src/constants/config.dart
@@ -8,6 +8,7 @@ import 'package:photo_manager/photo_manager.dart';
 import '../constants/typedefs.dart';
 import '../delegates/asset_picker_text_delegate.dart';
 import '../delegates/sort_path_delegate.dart';
+import '../models/special_item.dart';
 import 'constants.dart';
 import 'enums.dart';
 
@@ -29,8 +30,7 @@ class AssetPickerConfig {
     this.themeColor,
     this.pickerTheme,
     this.textDelegate,
-    this.specialItemPosition = SpecialItemPosition.none,
-    this.specialItemBuilder,
+    this.specialItems = const [],
     this.loadingIndicatorBuilder,
     this.selectPredicate,
     this.shouldRevertGrid,
@@ -56,13 +56,6 @@ class AssetPickerConfig {
               requestType == RequestType.common,
           'SpecialPickerType.wechatMoment and requestType '
           'cannot be set at the same time.',
-        ),
-        assert(
-          (specialItemBuilder == null &&
-                  identical(specialItemPosition, SpecialItemPosition.none)) ||
-              (specialItemBuilder != null &&
-                  !identical(specialItemPosition, SpecialItemPosition.none)),
-          'Custom item did not set properly.',
         );
 
   /// Selected assets.
@@ -168,13 +161,9 @@ class AssetPickerConfig {
 
   final AssetPickerTextDelegate? textDelegate;
 
-  /// Allow users set a special item in the picker with several positions.
-  /// 允许用户在选择器中添加一个自定义item，并指定位置
-  final SpecialItemPosition specialItemPosition;
-
-  /// The widget builder for the the special item.
-  /// 自定义item的构造方法
-  final SpecialItemBuilder<AssetPathEntity>? specialItemBuilder;
+  /// List of special items.
+  /// 自定义 item 列表
+  final List<SpecialItem<AssetPathEntity>> specialItems;
 
   /// Indicates the loading status for the builder.
   /// 指示目前加载的状态