+++ lib/src/delegates/asset_picker_builder_delegate.dart
@@ -22,6 +22,7 @@ import '../delegates/asset_grid_drag_selection_coordinator.dart';
 import '../delegates/asset_picker_text_delegate.dart';
 import '../internals/singleton.dart';
 import '../models/path_wrapper.dart';
+import '../models/special_item.dart';
 import '../provider/asset_picker_provider.dart';
 import '../widget/asset_picker.dart';
 import '../widget/asset_picker_app_bar.dart';
@@ -40,8 +41,7 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
     required this.initialPermission,
     this.gridCount = 4,
     this.pickerTheme,
-    this.specialItemPosition = SpecialItemPosition.none,
-    this.specialItemBuilder,
+    this.specialItems = const [],
     this.loadingIndicatorBuilder,
     this.selectPredicate,
     this.shouldRevertGrid,
@@ -89,13 +89,9 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
   /// 但某些情况下开发者需要亮色或自定义主题。
   final ThemeData? pickerTheme;
 
-  /// Allow users set a special item in the picker with several positions.
-  /// 允许用户在选择器中添加一个自定义 item，并指定位置
-  final SpecialItemPosition specialItemPosition;
-
-  /// The widget builder for the the special item.
-  /// 自定义 item 的构造方法
-  final SpecialItemBuilder<Path>? specialItemBuilder;
+  /// List of special items.
+  /// 自定义item列表
+  final List<SpecialItem<Path>> specialItems;
 
   /// Indicates the loading status for the builder.
   /// 指示目前加载的状态
@@ -175,12 +171,6 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
   /// Whether the picker is under the single asset mode.
   /// 选择器是否为单选模式
   bool get isSingleAssetMode;
-
-  /// Whether the delegate should build the special item.
-  /// 是否需要构建自定义 item
-  bool get shouldBuildSpecialItem =>
-      specialItemPosition != SpecialItemPosition.none &&
-      specialItemBuilder != null;
 
   /// Space between assets item widget.
   /// 资源部件之间的间隔
@@ -336,6 +326,7 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
   int? findChildIndexBuilder({
     required String id,
     required List<Asset> assets,
+    required List<SpecialItemFinalized> specialItemsFinalized,
     int placeholderCount = 0,
   }) =>
       null;
@@ -345,31 +336,49 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
   int assetsGridItemCount({
     required BuildContext context,
     required List<Asset> assets,
+    required List<SpecialItemFinalized> specialItemsFinalized,
     int placeholderCount = 0,
   });
+
+  List<SpecialItemFinalized> assetsGridSpecialItemsFinalized({
+    required BuildContext context,
+    required Path? path,
+  }) {
+    return specialItems
+        .map((item) {
+          final specialItem = item.builder?.call(
+            context,
+            path,
+            permissionNotifier.value,
+          );
+          if (specialItem != null) {
+            return SpecialItemFinalized(
+              position: item.position,
+              item: specialItem,
+            );
+          }
+          return null;
+        })
+        .nonNulls
+        .toList();
+  }
 
   /// Calculates the placeholder count in the assets grid.
   int assetsGridItemPlaceholderCount({
     required BuildContext context,
     required PathWrapper<Path>? pathWrapper,
     required bool onlyOneScreen,
+    required List<SpecialItemFinalized> specialItemsFinalized,
   }) {
     if (onlyOneScreen) {
       return 0;
     }
+
     final bool gridRevert = effectiveShouldRevertGrid(context);
     int totalCount = pathWrapper?.assetCount ?? 0;
-    // If user chose a special item's position, add 1 count.
-    if (specialItemPosition != SpecialItemPosition.none) {
-      final specialItem = specialItemBuilder?.call(
-        context,
-        pathWrapper?.path,
-        totalCount,
-      );
-      if (specialItem != null) {
-        totalCount += 1;
-      }
-    }
+    // Add special items' count.
+    totalCount += specialItemsFinalized.length;
+
     final int result;
     if (gridRevert && totalCount % gridCount != 0) {
       // When there are left items that not filled into one row,
@@ -379,6 +388,7 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
       // Otherwise, we don't need placeholders.
       result = 0;
     }
+
     return result;
   }
 
@@ -387,19 +397,12 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
     required BuildContext context,
     required BoxConstraints constraints,
     required PathWrapper<Path>? pathWrapper,
+    required List<SpecialItemFinalized> specialItemsFinalized,
   }) {
     int totalCount = pathWrapper?.assetCount ?? 0;
-    // If user chose a special item's position, add 1 count.
-    if (specialItemPosition != SpecialItemPosition.none) {
-      final specialItem = specialItemBuilder?.call(
-        context,
-        pathWrapper?.path,
-        totalCount,
-      );
-      if (specialItem != null) {
-        totalCount += 1;
-      }
-    }
+    // Add special items' count.
+    totalCount += specialItemsFinalized.length;
+
     // Here we got a magic calculation. [itemSpacing] needs to be divided by
     // [gridCount] since every grid item is squeezed by the [itemSpacing],
     // and it's actual size is reduced with [itemSpacing / gridCount].
@@ -427,11 +430,12 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
 
   /// The item builder for the assets' grid.
   /// 资源列表项的构建
-  Widget assetGridItemBuilder(
-    BuildContext context,
-    int index,
-    List<Asset> currentAssets,
-  );
+  Widget assetGridItemBuilder({
+    required BuildContext context,
+    required int index,
+    required List<Asset> currentAssets,
+    required List<SpecialItemFinalized> specialItemsFinalized,
+  });
 
   /// The [Semantics] builder for the assets' grid.
   /// 资源列表项的语义构建
@@ -440,6 +444,7 @@ abstract class AssetPickerBuilderDelegate<Asset, Path> {
     int index,
     Asset asset,
     Widget child,
+    List<SpecialItemFinalized> specialItemsFinalized,
   );
 
   /// The item builder for audio type of asset.
@@ -826,8 +831,7 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
     required super.initialPermission,
     super.gridCount,
     super.pickerTheme,
-    super.specialItemPosition,
-    super.specialItemBuilder,
+    super.specialItems = const [],
     super.loadingIndicatorBuilder,
     super.selectPredicate,
     super.shouldRevertGrid,
@@ -1227,9 +1231,14 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
     return AssetPickerAppBarWrapper(
       appBar: appBar(context),
       body: Consumer<T>(
-        builder: (BuildContext context, T p, _) {
-          final bool shouldDisplayAssets =
-              p.hasAssetsToDisplay || shouldBuildSpecialItem;
+        builder: (context, p, _) {
+          final hasAssetsToDisplay = p.hasAssetsToDisplay;
+          final shouldBuildSpecialItems = assetsGridSpecialItemsFinalized(
+            context: context,
+            path: p.currentPath?.path,
+          ).isNotEmpty;
+          final shouldDisplayAssets =
+              hasAssetsToDisplay || shouldBuildSpecialItems;
           return AnimatedSwitcher(
             duration: switchingPathDuration,
             child: shouldDisplayAssets
@@ -1278,10 +1287,15 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
         children: <Widget>[
           Positioned.fill(
             child: Consumer<T>(
-              builder: (_, p, __) {
+              builder: (context, p, _) {
+                final hasAssetsToDisplay = p.hasAssetsToDisplay;
+                final shouldBuildSpecialItems = assetsGridSpecialItemsFinalized(
+                  context: context,
+                  path: p.currentPath?.path,
+                ).isNotEmpty;
+                final shouldDisplayAssets =
+                    hasAssetsToDisplay || shouldBuildSpecialItems;
                 final Widget child;
-                final bool shouldDisplayAssets =
-                    p.hasAssetsToDisplay || shouldBuildSpecialItem;
                 if (shouldDisplayAssets) {
                   child = Stack(
                     children: <Widget>[
@@ -1344,21 +1358,13 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
       builder: (context, wrapper, _) {
         // First, we need the count of the assets.
         int totalCount = wrapper?.assetCount ?? 0;
-        final Widget? specialItem;
-        // If user chose a special item's position, add 1 count.
-        if (specialItemPosition != SpecialItemPosition.none) {
-          specialItem = specialItemBuilder?.call(
-            context,
-            wrapper?.path,
-            totalCount,
-          );
-          if (specialItem != null) {
-            totalCount += 1;
-          }
-        } else {
-          specialItem = null;
-        }
-        if (totalCount == 0 && specialItem == null) {
+        final specialItemsFinalized = assetsGridSpecialItemsFinalized(
+          context: context,
+          path: wrapper?.path,
+        );
+        totalCount += specialItemsFinalized.length;
+
+        if (totalCount == 0 && specialItemsFinalized.isEmpty) {
           return loadingIndicator(context);
         }
 
@@ -1377,6 +1383,7 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
             context: context,
             pathWrapper: wrapper,
             onlyOneScreen: onlyOneScreen,
+            specialItemsFinalized: specialItemsFinalized,
           );
           return SliverGrid(
             delegate: SliverChildBuilderDelegate(
@@ -1389,10 +1396,10 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
                 }
 
                 Widget child = assetGridItemBuilder(
-                  context,
-                  index,
-                  assets,
-                  specialItem: specialItem,
+                  context: context,
+                  index: index,
+                  currentAssets: assets,
+                  specialItemsFinalized: specialItemsFinalized,
                 );
 
                 // Enables drag-to-select when:
@@ -1483,7 +1490,7 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
                 context: context,
                 assets: assets,
                 placeholderCount: placeholderCount,
-                specialItem: specialItem,
+                specialItemsFinalized: specialItemsFinalized,
               ),
               findChildIndexCallback: (Key? key) {
                 if (key is ValueKey<String>) {
@@ -1491,6 +1498,7 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
                     id: key.value,
                     assets: assets,
                     placeholderCount: placeholderCount,
+                    specialItemsFinalized: specialItemsFinalized,
                   );
                 }
                 return null;
@@ -1523,6 +1531,7 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
               context: context,
               constraints: constraints,
               pathWrapper: wrapper,
+              specialItemsFinalized: specialItemsFinalized,
             );
 
             final reverted = gridRevert && !onlyOneScreen;
@@ -1585,32 +1594,41 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
   ///      图片和视频类型
   ///  * 在索引到达倒数第三列的时候加载更多资源。
   @override
-  Widget assetGridItemBuilder(
-    BuildContext context,
-    int index,
-    List<AssetEntity> currentAssets, {
-    Widget? specialItem,
+  Widget assetGridItemBuilder({
+    required BuildContext context,
+    required int index,
+    required List<AssetEntity> currentAssets,
+    required List<SpecialItemFinalized> specialItemsFinalized,
   }) {
     final p = context.read<T>();
     final int length = currentAssets.length;
     final PathWrapper<AssetPathEntity>? currentWrapper = p.currentPath;
     final AssetPathEntity? currentPathEntity = currentWrapper?.path;
 
-    if (specialItem != null) {
-      if ((index == 0 && specialItemPosition == SpecialItemPosition.prepend) ||
-          (index == length &&
-              specialItemPosition == SpecialItemPosition.append)) {
-        return specialItem;
+    final prependItems = <SpecialItemFinalized>[];
+    final appendItems = <SpecialItemFinalized>[];
+    for (final model in specialItemsFinalized) {
+      switch (model.position) {
+        case SpecialItemPosition.prepend:
+          prependItems.add(model);
+        case SpecialItemPosition.append:
+          appendItems.add(model);
       }
     }
 
-    final int currentIndex;
-    if (specialItem != null &&
-        specialItemPosition == SpecialItemPosition.prepend) {
-      currentIndex = index - 1;
-    } else {
-      currentIndex = index;
+    if (prependItems.isNotEmpty) {
+      if (index < prependItems.length) {
+        return specialItemsFinalized[index].item;
+      }
     }
+
+    if (appendItems.isNotEmpty) {
+      if (index >= length + prependItems.length) {
+        return specialItemsFinalized[index - length].item;
+      }
+    }
+
+    final currentIndex = index - prependItems.length;
 
     if (currentPathEntity == null) {
       return const SizedBox.shrink();
@@ -1641,14 +1659,23 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
         itemBannedIndicator(context, asset),
       ],
     );
-    return assetGridItemSemanticsBuilder(context, index, asset, content);
+    return assetGridItemSemanticsBuilder(
+      context,
+      index,
+      asset,
+      content,
+      specialItemsFinalized,
+    );
   }
 
-  int semanticIndex(int index) {
-    if (specialItemPosition != SpecialItemPosition.prepend) {
-      return index + 1;
-    }
-    return index;
+  int assetGridItemSemanticIndex(
+    int index,
+    List<SpecialItemFinalized> specialItemsFinalized,
+  ) {
+    final prependItems = specialItemsFinalized.where(
+      (model) => model.position == SpecialItemPosition.prepend,
+    );
+    return index - prependItems.length;
   }
 
   @override
@@ -1657,6 +1684,7 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
     int index,
     AssetEntity asset,
     Widget child,
+    List<SpecialItemFinalized> specialItemsFinalized,
   ) {
     return ValueListenableBuilder<bool>(
       valueListenable: isSwitchingPath,
@@ -1674,7 +1702,7 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
             final int selectedIndex = p.selectedAssets.indexOf(asset) + 1;
             final labels = <String>[
               '${semanticsTextDelegate.semanticTypeLabel(asset.type)}'
-                  '${semanticIndex(index)}',
+                  '${assetGridItemSemanticIndex(index, specialItemsFinalized)}',
               asset.createDateTime.toString().replaceAll('.000', ''),
               if (asset.type == AssetType.audio ||
                   asset.type == AssetType.video)
@@ -1704,7 +1732,8 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
               onLongPressHint: semanticsTextDelegate.sActionPreviewHint,
               selected: isSelected,
               sortKey: OrdinalSortKey(
-                semanticIndex(index).toDouble(),
+                assetGridItemSemanticIndex(index, specialItemsFinalized)
+                    .toDouble(),
                 name: 'GridItem',
               ),
               value: selectedIndex > 0 ? '$selectedIndex' : null,
@@ -1717,7 +1746,8 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
                       }
                     : null,
                 child: IndexedSemantics(
-                  index: semanticIndex(index),
+                  index:
+                      assetGridItemSemanticIndex(index, specialItemsFinalized),
                   child: child,
                 ),
               ),
@@ -1733,12 +1763,14 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
   int findChildIndexBuilder({
     required String id,
     required List<AssetEntity> assets,
+    required List<SpecialItemFinalized> specialItemsFinalized,
     int placeholderCount = 0,
   }) {
+    final prependItems = specialItemsFinalized.where(
+      (model) => model.position == SpecialItemPosition.prepend,
+    );
     int index = assets.indexWhere((AssetEntity e) => e.id == id);
-    if (specialItemPosition == SpecialItemPosition.prepend) {
-      index += 1;
-    }
+    index += prependItems.length;
     index += placeholderCount;
     return index;
   }
@@ -1747,30 +1779,10 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
   int assetsGridItemCount({
     required BuildContext context,
     required List<AssetEntity> assets,
+    required List<SpecialItemFinalized> specialItemsFinalized,
     int placeholderCount = 0,
-    Widget? specialItem,
   }) {
-    final PathWrapper<AssetPathEntity>? currentWrapper =
-        context.select<T, PathWrapper<AssetPathEntity>?>(
-      (T p) => p.currentPath,
-    );
-    final AssetPathEntity? currentPathEntity = currentWrapper?.path;
-    final int length = assets.length + placeholderCount;
-
-    // Return 1 if the [specialItem] build something.
-    if (currentPathEntity == null && specialItem != null) {
-      return placeholderCount + 1;
-    }
-
-    // Return actual length if the current path is all.
-    // 如果当前目录是全部内容，则返回实际的内容数量。
-    if (currentPathEntity?.isAll != true && specialItem == null) {
-      return length;
-    }
-    return switch (specialItemPosition) {
-      SpecialItemPosition.none => length,
-      SpecialItemPosition.prepend || SpecialItemPosition.append => length + 1,
-    };
+    return assets.length + specialItemsFinalized.length + placeholderCount;
   }
 
   @override
@@ -2055,11 +2067,8 @@ class DefaultAssetPickerBuilderDelegate<T extends DefaultAssetPickerProvider>
               child: Selector<T, List<PathWrapper<AssetPathEntity>>>(
                 selector: (_, T p) => p.paths,
                 builder: (_, List<PathWrapper<AssetPathEntity>> paths, __) {
-                  final List<PathWrapper<AssetPathEntity>> filtered = paths
-                      .where(
-                        (PathWrapper<AssetPathEntity> p) => p.assetCount != 0,
-                      )
-                      .toList();
+                  final filtered =
+                      paths.where((p) => p.assetCount != 0).toList();
                   return ListView.separated(
                     padding: const EdgeInsetsDirectional.only(top: 1),
                     shrinkWrap: true,